requirements.txt
beautifulsoup4==4.6.0
bleach==2.1.3
bs4==0.0.1
cachetools==2.0.1
certifi==2018.1.18
chardet==3.0.4
click==6.7
colorama==0.3.7
cycler==0.10.0
cymem==1.31.2
cytoolz==0.8.2
decorator==4.2.1
dill==0.2.7.1
docutils==0.14
entrypoints==0.2.3
feedparser==5.2.1
future==0.16.0
geocoder==1.38.1
geographiclib==1.49
geopy==1.13.0
google-api-core==1.1.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-firestore==0.29.0
googleapis-common-protos==1.5.3
grpcio==1.10.1
html5lib==1.0.1
idna==2.6
itsdangerous==0.24
jedi==0.11.1
jmespath==0.9.3
jsonschema==2.6.0
kiwisolver==1.0.1
MarkupSafe==1.0
mistune==0.8.3
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
numpy==1.14.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pathlib==1.0.1
pexpect==4.4.0
pickleshare==0.7.4
plac==0.9.6
preshed==1.0.0
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
pymongo==3.6.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2018.4
PyYAML==3.12
pyzmq==17.0.0
qtconsole==4.3.1
ratelim==0.1.6
regex==2017.4.5
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.0.1
selenium==3.11.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sklearn==0.0
spacy==2.0.11
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.2
toolz==0.9.0
tqdm==4.22.0
traitlets==4.3.2
ujson==1.35
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.1.4
word2number==1.1
wrapt==1.10.11
